+++ platform/fabric/core/generic/finality/fabric.go
@@ -13,11 +13,14 @@ import (
 	"github.com/hyperledger-labs/fabric-smart-client/platform/fabric/core/generic/delivery"
 	"github.com/hyperledger-labs/fabric-smart-client/platform/fabric/core/generic/peer"
 	"github.com/hyperledger-labs/fabric-smart-client/platform/fabric/driver"
+	"github.com/hyperledger-labs/fabric-smart-client/platform/view/services/flogging"
 	"github.com/hyperledger-labs/fabric-smart-client/platform/view/services/grpc"
 	ab "github.com/hyperledger/fabric-protos-go/orderer"
 	"github.com/pkg/errors"
 	"go.uber.org/zap/zapcore"
 )
+
+var logger = flogging.MustGetLogger("fabric-sdk.core")
 
 type PeerManager interface {
 	NewPeerClientForAddress(cc grpc.ConnectionConfig) (peer.Client, error)
