--- platform/fabric/core/generic/txconfig.go
@@ -1,12 +1,0 @@
-/*
-Copyright IBM Corp. All Rights Reserved.
-
-SPDX-License-Identifier: Apache-2.0
-*/
-
-package generic
-
-const (
-	channelConfigKey = "CHANNEL_CONFIG_ENV_BYTES"
-	peerNamespace    = "_configtx"
-)
